Fix darwin arm64 cross build
Cast syscall device numbers at the proc locks boundary and add a darwin/arm64 cross-build target to make verify so Mac build regressions are caught before publishing.

Validation: GOOS=darwin GOARCH=arm64 go build -o bin/cdp-darwin-arm64 ./cmd/cdp; make verify; make install; make e2e-installed.

+++ internal/cli/cron_commands.go
@@ -1285,7 +1285,7 @@ func cronFlockOwnerFromProcLocks(path string) (cronFlockOwner, bool) {
 	if !ok {
 		return cronFlockOwner{}, false
 	}
-	device := procLocksDevice(stat.Dev)
+	device := procLocksDevice(uint64(stat.Dev))
 	inode := strconv.FormatUint(stat.Ino, 10)
 	file, err := os.Open("/proc/locks")
 	if err != nil {
